requirements/base.txt
# Bleeding edge Django

django==1.11.1

# Configuration
django-environ==0.4.3
django-secure==1.0.1

# Forms
django-braces==1.11.0
django-crispy-forms==1.6.1
django-floppyforms==1.7.0

# Models
django-model-utils==2.5.2

# Images
Pillow==2.9.0

# For authentication with OSM
social-auth-app-django==1.2.0
social-auth-core==1.3.0
oauth2==1.9.0.post1

# DRF and DRF-gis
djangorestframework==3.6.3
djangorestframework-gis==0.11.2
djangorestframework-csv==2.0.0
django-filter==1.0.4
django-rest-swagger==2.1.2
django-cors-headers==2.1.0
# coreapi==2.3.0

# For the persistence stores
psycopg2==2.7.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2017.2

# Celery support
celery==3.1.25

# Redis support
django-redis==4.8.0
redis>=2.10.5

# Cache
django-cachalot==1.5.0

# Your custom requirements go here
PyYAML==3.12
# using the premerge version of osmcha
osmcha==0.3.8

# git python is required by the frontend management command
GitPython==2.1.5

